tests/test_model.py
import unittest
import mlflow
import mlflow.pyfunc
from mlflow.tracking import MlflowClient
import os
import pandas as pd
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
import pickle
from dotenv import load_dotenv

class TestModelLoading(unittest.TestCase):

    @classmethod
    def setUpClass(cls):
        # Load environment variables
        load_dotenv()
        
        # Set up DagsHub credentials for MLflow tracking
        dagshub_token = os.getenv("DAGSHUB_PAT")
        if not dagshub_token:
            raise EnvironmentError("DAGSHUB_PAT environment variable is not set")

        os.environ["MLFLOW_TRACKING_USERNAME"] = dagshub_token
        os.environ["MLFLOW_TRACKING_PASSWORD"] = dagshub_token

        dagshub_url = "https://dagshub.com"
        repo_owner = "DeepuML"
        repo_name = "Ml-OPS-Project-2"

        # Set up MLflow tracking URI
        mlflow.set_tracking_uri(f'{dagshub_url}/{repo_owner}/{repo_name}.mlflow')

        # Load the new model from MLflow model registry
        cls.new_model_name = "my_model"
        cls.new_model_version = cls.get_latest_model_version(cls.new_model_name)
        
        if cls.new_model_version:
            cls.new_model_uri = f'models:/{cls.new_model_name}/{cls.new_model_version}'
            try:
                cls.new_model = mlflow.pyfunc.load_model(cls.new_model_uri)
            except Exception as e:
                print(f"Warning: Could not load model from registry: {e}")
                # Fallback to loading from local files
                cls.new_model = pickle.load(open('models/model.pkl', 'rb'))
        else:
            print("Warning: No model found in registry, loading from local files")
            cls.new_model = pickle.load(open('models/model.pkl', 'rb'))

        # Load the vectorizer
        cls.vectorizer = pickle.load(open('models/vectorizer.pkl', 'rb'))

        # Load holdout test data
        cls.holdout_data = pd.read_csv('data/processed/test_bow.csv')

    @staticmethod
    def get_latest_model_version(model_name):
        """Get the latest version of a registered model."""
        try:
            client = MlflowClient()
            # Get all versions of the model and return the latest
            model_versions = client.search_model_versions(f"name='{model_name}'")
            if model_versions:
                # Sort by version number and get the latest
                latest_version = max(model_versions, key=lambda x: int(x.version))
                return latest_version.version
            return None
        except Exception as e:
            print(f"Warning: Could not get latest model version: {e}")
            return "1"  # Default to version 1 if we can't get the latest

    def test_model_loaded_properly(self):
        self.assertIsNotNone(self.new_model)

    def test_model_signature(self):
        # Create a dummy input for the model based on expected input shape
        input_text = "hi how are you"
        input_data = self.vectorizer.transform([input_text])
        input_df = pd.DataFrame(input_data.toarray(), columns=[str(i) for i in range(input_data.shape[1])])

        # Predict using the new model to verify the input and output shapes
        if hasattr(self.new_model, 'predict'):
            # For both MLflow and sklearn models
            if str(type(self.new_model)).find('mlflow') != -1:
                # MLflow model - expects DataFrame
                prediction = self.new_model.predict(input_df)
            else:
                # Sklearn model - can use array
                prediction = self.new_model.predict(input_data.toarray())
        else:
            self.fail("Model does not have predict method")

        # Verify the input shape
        self.assertEqual(input_df.shape[1], len(self.vectorizer.get_feature_names_out()))

        # Verify the output shape (assuming binary classification with a single output)
        self.assertGreater(len(prediction), 0)
        self.assertTrue(isinstance(prediction, (list, tuple)) or hasattr(prediction, 'shape'))

    def test_model_performance(self):
        # Extract features and labels from holdout test data
        X_holdout = self.holdout_data.iloc[:,0:-1]
        y_holdout = self.holdout_data.iloc[:,-1]

        # Predict using the new model
        if str(type(self.new_model)).find('mlflow') != -1:
            # MLflow model - expects DataFrame
            y_pred_new = self.new_model.predict(X_holdout)
        else:
            # Sklearn model - can use array
            y_pred_new = self.new_model.predict(X_holdout.values)

        # Calculate performance metrics for the new model
        accuracy_new = accuracy_score(y_holdout, y_pred_new)
        precision_new = precision_score(y_holdout, y_pred_new, average='weighted', zero_division=0)
        recall_new = recall_score(y_holdout, y_pred_new, average='weighted', zero_division=0)
        f1_new = f1_score(y_holdout, y_pred_new, average='weighted', zero_division=0)

        # Print metrics for debugging
        print(f"Model Performance Metrics:")
        print(f"Accuracy: {accuracy_new:.4f}")
        print(f"Precision: {precision_new:.4f}")
        print(f"Recall: {recall_new:.4f}")
        print(f"F1 Score: {f1_new:.4f}")

        # Define expected thresholds for the performance metrics (more realistic)
        expected_accuracy = 0.60
        expected_precision = 0.60
        expected_recall = 0.60
        expected_f1 = 0.60

        # Assert that the new model meets the performance thresholds
        self.assertGreaterEqual(accuracy_new, expected_accuracy, f'Accuracy {accuracy_new:.4f} should be at least {expected_accuracy}')
        self.assertGreaterEqual(precision_new, expected_precision, f'Precision {precision_new:.4f} should be at least {expected_precision}')
        self.assertGreaterEqual(recall_new, expected_recall, f'Recall {recall_new:.4f} should be at least {expected_recall}')
        self.assertGreaterEqual(f1_new, expected_f1, f'F1 score {f1_new:.4f} should be at least {expected_f1}')

if __name__ == "__main__":
    unittest.main()